chore: Firebase 기능 쪼개기
