+++ alsongDalsong/ASNetworkKit/ASNetworkKit/Firebase/ASFirebaseDatabase.swift
@@ -1,0 +1,37 @@
+import ASEntity
+import Combine
+@preconcurrency internal import FirebaseFirestore
+
+public final class ASFirebaseDatabase: ASFirebaseDatabaseProtocol {
+    private let firestoreRef = Firestore.firestore()
+    private var roomListeners: ListenerRegistration?
+    
+    private var roomPublisher = PassthroughSubject<Room, Error>()
+    
+    public func addRoomListener(roomNumber: String) -> AnyPublisher<Room, Error> {
+        let roomRef = firestoreRef.collection("rooms").document(roomNumber)
+        let listener = roomRef.addSnapshotListener { documentSnapshot, error in
+            if let error {
+                return
+            }
+            
+            guard let document = documentSnapshot, document.exists else {
+                return self.roomPublisher.send(completion: .failure(ASNetworkErrors.FirebaseListenerError))
+            }
+            
+            do {
+                let room = try document.data(as: Room.self)
+                return self.roomPublisher.send(room)
+            } catch {
+                return self.roomPublisher.send(completion: .failure(ASNetworkErrors.FirebaseListenerError))
+            }
+        }
+        
+        roomListeners = listener
+        return roomPublisher.eraseToAnyPublisher()
+    }
+    
+    public func removeRoomListener(roomNumber: String) {
+        roomListeners?.remove()
+    }
+}
