+++ alsongDalsong/ASNetworkKit/ASNetworkKit/Protocols/ASFirebaseStorageProtocol.swift
